Modified circulation api

+++ circulation/serializers.py
@@ -4,11 +4,12 @@ from circulation.models import LaundryCirculation, InsideLaundry
 
 
 class InsideLaundrySerializer(serializers.ModelSerializer):
+
     class Meta:
         from .models import InsideLaundry
 
         model = InsideLaundry
-        fields = ('id', 'check_in_date', 'card_no', 'clothes')
+        fields = ('check_in_date', 'card_no', 'clothes', 'ready')
 
 
 class LaundryCirculationSerializer(serializers.ModelSerializer):
@@ -16,12 +17,21 @@ class LaundryCirculationSerializer(serializers.ModelSerializer):
     def create(self, validated_data):
 
         card_no = validated_data.pop('card_no')
+        ready = InsideLaundry.objects.get(card_no=card_no).ready
 
-        InsideLaundry.objects.get(card_no=card_no).delete()
+        if ready:
+            check_in_date = InsideLaundry.objects.get(card_no=card_no).check_in_date
+            clothes = InsideLaundry.objects.get(card_no=card_no).clothes
 
-        return LaundryCirculation.objects.create(
-            card_no=card_no,
-        )
+            InsideLaundry.objects.get(card_no=card_no).delete()
+
+            return LaundryCirculation.objects.create(
+                card_no=card_no,
+                check_in_date=check_in_date,
+                clothes=clothes,
+            )
+        else:
+            return Exception('The bag is not yet ready.')
 
     class Meta:
         model = LaundryCirculation
@@ -30,6 +40,7 @@ class LaundryCirculationSerializer(serializers.ModelSerializer):
 
 
 class LaundryListSerializer(serializers.ModelSerializer):
+
     class Meta:
         from .models import LaundryCirculation
